# More efficient implementation of digit counting during serialization of numbers