## Before
#### Corretto 8
```
[info] Benchmark                                 (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfIntsWriting.jsoniterScala             128  thrpt   25  718466.540 ± 23233.181  ops/s
[info] ArrayOfIntsWriting.jsoniterScalaPrealloc     128  thrpt   25  731960.156 ± 13610.345  ops/s
```
#### Corretto 11
```
[info] Benchmark                                 (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfIntsWriting.jsoniterScala             128  thrpt   25  802355.122 ± 52639.382  ops/s
[info] ArrayOfIntsWriting.jsoniterScalaPrealloc     128  thrpt   25  880545.903 ± 15563.172  ops/s
```
#### OpenJDK 17
```
[info] Benchmark                                 (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfIntsWriting.jsoniterScala             128  thrpt   25  900899.758 ± 32470.493  ops/s
[info] ArrayOfIntsWriting.jsoniterScalaPrealloc     128  thrpt   25  923576.891 ± 18464.388  ops/s
```
## After
#### Corretto 8
```
[info] Benchmark                                 (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfIntsWriting.jsoniterScala             128  thrpt   25  786654.818 ± 12677.207  ops/s
[info] ArrayOfIntsWriting.jsoniterScalaPrealloc     128  thrpt   25  803984.994 ± 11676.662  ops/s
```
#### Corretto 11
```
[info] Benchmark                                 (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfIntsWriting.jsoniterScala             128  thrpt   25  857944.977 ± 44241.449  ops/s
[info] ArrayOfIntsWriting.jsoniterScalaPrealloc     128  thrpt   25  919552.479 ± 21334.477  ops/s
```
#### OpenJDK 17
```
[info] Benchmark                                 (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfIntsWriting.jsoniterScala             128  thrpt   25  933313.338 ± 31055.176  ops/s
[info] ArrayOfIntsWriting.jsoniterScalaPrealloc     128  thrpt   25  935877.114 ± 29157.245  ops/s
```